Make explore observations layout persistent; closes #1052
